Create category explorer screen to simplify routing logic
